requirements.txt
fastapi
uvicorn
scikit-learn
pyyaml
requests
